+++ src/radiolibs.hpp
@@ -95,7 +95,7 @@ public:
     rig_port_e convGkConnTypeToHamlib(const GekkoFyre::AmateurRadio::GkConnType &conn_type);
 
     void gkInitRadioRig(std::shared_ptr<AmateurRadio::Control::GkRadio> radio_ptr, std::shared_ptr<Database::Settings::GkUsbPort> usb_ptr);
-    float calibrateAudioInputSignal(const float *data_buf);
+    int16_t calibrateAudioInputSignal(const int16_t *data_buf);
 
     libusb_context *initUsbLib();
     QMap<std::string, Database::Settings::GkUsbPort> enumUsbDevices(libusb_context *usb_ctx_ptr);
